Extract shared cover scoring module, parse OCR page-type tags (#977)
The cover scorer now parses <page-type> tags from OCR text when
page.page_type is null (which is most pages). Also adds detection for
bookplates ("library of", "botanical garden", "book fund"), endpapers
("textured surface", "inside cover", "marbled"), Google Books patterns,
and a title-match bonus that boosts pages containing the book's title.

Extracts duplicated scoring logic from pipeline-orchestrator.mjs and
smart-cover-selection.mjs into scripts/lib/cover-scoring.mjs.
