+++ scripts/maintenance/smart-cover-selection.mjs
@@ -3,6 +3,7 @@
  *
  * Picks the best cover by analyzing OCR text content — not just page_type,
  * but the actual descriptions, meta tags, and transcribed text.
+ * Parses <page-type> tags from OCR when page_type field is null.
  *
  * Cost: FREE
  *
@@ -16,6 +17,7 @@
  */
 
 import { MongoClient } from 'mongodb';
+import { scorePageForCover } from '../lib/cover-scoring.mjs';
 
 const DRY_RUN = process.argv.includes('--dry-run');
 const SKIP_MANUAL = process.argv.includes('--skip-manual');
@@ -35,160 +37,13 @@ const PROVIDER = (() => {
 })();
 
 const SCAN_PAGES = 20;
-const BATCH_SIZE = 50; // smaller batches since we fetch more OCR data
+const BATCH_SIZE = 50;
 
 const client = new MongoClient(process.env.MONGODB_URI, {
   serverSelectionTimeoutMS: 30000, connectTimeoutMS: 30000, socketTimeoutMS: 60000,
 });
 await client.connect();
 const db = client.db('bookstore');
-
-/**
- * Score a page as a potential cover based on OCR content analysis.
- * Returns { score, reason }.
- */
-function scorePage(page) {
-  const ocr = (page.ocr_text || '').toLowerCase();
-  const pageType = page.page_type;
-  const pageNum = page.page_number;
-  let score = 0;
-  let reason = pageType || 'unknown';
-
-  // === NEGATIVE signals: things that make terrible covers ===
-
-  // Blank pages
-  if (pageType === 'blank' && !ocr.includes('title')) return { score: -100, reason: 'blank' };
-
-  // Book binding / cover photos (external shots of the physical book)
-  if (ocr.includes('front cover') || ocr.includes('external') && ocr.includes('cover') ||
-      ocr.includes('binding') && ocr.includes('spine') || ocr.includes('fore-edge') ||
-      ocr.includes('closed book') || ocr.includes('book block') ||
-      ocr.includes('leather-bound') || ocr.includes('leather bound') ||
-      (ocr.includes('blind-stamp') || ocr.includes('blind stamp')) && ocr.includes('cover')) {
-    return { score: -80, reason: 'physical book photo' };
-  }
-
-  // Digitizer inserts and digital portal pages
-  if (ocr.includes('digitized by') || ocr.includes('google books') ||
-      ocr.includes('internet archive') && ocr.includes('logo') ||
-      ocr.includes('proquest') || ocr.includes('early european books') ||
-      ocr.includes('hathitrust') || ocr.includes('scan sheet') ||
-      ocr.includes('e-rara.ch') || ocr.includes('www.e-rara') ||
-      ocr.includes('gallica.bnf') || ocr.includes('mdz-nbn') ||
-      ocr.includes('digital.staatsbibliothek') || ocr.includes('daten.digitale-sammlungen')) {
-    return { score: -70, reason: 'digitizer insert' };
-  }
-
-  // BPH pelican bookplate (Embassy of the Free Mind / Bibliotheca Philosophica Hermetica)
-  const isBphBookplate = (ocr.includes('pelican') && (ocr.includes('piety') || ocr.includes('nest') || ocr.includes('young') || ocr.includes('hermetica'))) ||
-      ocr.includes('philosophia hermetica') || ocr.includes('philosophica hermetica');
-  if (isBphBookplate && pageType !== 'title-page') {
-    return { score: -65, reason: 'BPH pelican bookplate' };
-  }
-
-  // Ex-libris / bookplates — but only if the page is PRIMARILY a bookplate,
-  // not a title page that happens to have a small ownership inscription
-  const hasExLibris = ocr.includes('ex-libris') || ocr.includes('exlibris') || ocr.includes('ex libris') ||
-      ocr.includes('bookplate') || (ocr.includes('ownership') && ocr.includes('stamp')) ||
-      ocr.includes('library stamp') || ocr.includes('library sticker');
-  if (hasExLibris && pageType !== 'title-page') {
-    return { score: -60, reason: 'ex-libris/bookplate' };
-  }
-
-  // Bleed-through / ghosting pages
-  if (ocr.includes('bleed-through') || ocr.includes('ghosting') ||
-      ocr.includes('mirrored impression')) {
-    return { score: -50, reason: 'bleed-through' };
-  }
-
-  // Tables of contents, indices
-  if (pageType === 'toc' || pageType === 'index') return { score: -20, reason: pageType };
-
-  // Colophon
-  if (pageType === 'colophon') return { score: -10, reason: 'colophon' };
-
-  // === POSITIVE signals: things that make great covers ===
-
-  // Title pages with actual title text (centered headings in OCR)
-  const hasHeadings = (ocr.match(/^#+ .+/gm) || []).length;
-  const isTitlePage = pageType === 'title-page';
-  const looksLikeTitlePage = !pageType && hasHeadings >= 3;
-
-  if (isTitlePage) {
-    score += 80;
-    reason = 'title-page';
-  } else if (looksLikeTitlePage) {
-    score += 70;
-    reason = 'likely title-page (headings)';
-  }
-
-  if (isTitlePage || looksLikeTitlePage) {
-    // Bonus: publisher marks / imprints (Latin publishing terms)
-    if (ocr.includes('excudebat') || ocr.includes('typis') || ocr.includes('apud') ||
-        ocr.includes('impensis') || ocr.includes('sumptibus') || ocr.includes('officina') ||
-        ocr.includes('printed by') || ocr.includes('published by') || ocr.includes('printed for')) {
-      score += 15;
-      reason = (isTitlePage ? 'title-page' : 'likely title-page') + ' with imprint';
-    }
-
-    // Bonus: decorative elements described in OCR
-    if (ocr.includes('decorative') || ocr.includes('ornamental') || ocr.includes('border') ||
-        ocr.includes('frame') || ocr.includes('vignette') || ocr.includes('woodcut') ||
-        ocr.includes('architectural') || ocr.includes('printer\'s mark') ||
-        ocr.includes('printer\'s device') || ocr.includes('colophon device')) {
-      score += 20;
-      reason = 'decorated title-page';
-    }
-  }
-
-  // Frontispieces — but distinguish real frontispieces from binding photos
-  if (pageType === 'frontispiece') {
-    // Check if it's actually a physical cover photo mislabeled as frontispiece
-    if (ocr.includes('cover') && (ocr.includes('leather') || ocr.includes('binding') || ocr.includes('marbled'))) {
-      return { score: -80, reason: 'binding photo (mislabeled frontispiece)' };
-    }
-
-    score += 90;
-    reason = 'frontispiece';
-
-    // Bonus for engravings / allegorical imagery
-    if (ocr.includes('engrav') || ocr.includes('allegor') || ocr.includes('portrait') ||
-        ocr.includes('emblem') || ocr.includes('woodcut')) {
-      score += 15;
-      reason = 'frontispiece with engraving';
-    }
-  }
-
-  // Illustrations
-  if (pageType === 'illustration') {
-    // Skip if it's a photo of the physical book
-    if (ocr.includes('photograph') && (ocr.includes('fore-edge') || ocr.includes('book') && ocr.includes('closed'))) {
-      return { score: -80, reason: 'physical book photo' };
-    }
-    score += 60;
-    reason = 'illustration';
-  }
-
-  // Dedication pages — sometimes nice but not ideal
-  if (pageType === 'dedication') {
-    score += 15;
-    reason = 'dedication';
-  }
-
-  // Plain text — lowest priority
-  if (pageType === 'text' && score === 0) {
-    score += 5;
-    reason = 'text';
-  }
-
-  // === Position bonuses ===
-  // Slight preference for earlier pages (covers are usually in first 10)
-  if (pageNum <= 5) score += 5;
-  else if (pageNum <= 10) score += 3;
-  else if (pageNum > 15) score -= 3;
-
-  return { score, reason };
-}
 
 function getPageImageUrl(page) {
   const isUsable = (u) => u && (u.startsWith('http://') || u.startsWith('https://'));
@@ -249,11 +104,9 @@ for (let i = 0; i < allBooks.length; i += BATCH_SIZE) {
     .sort({ book_id: 1, page_number: 1 })
     .toArray();
 
-  // Extract OCR text (first 800 chars for scoring) and group by book
+  // Group by book — keep ocr.data intact for the shared scorer
   const pagesByBook = new Map();
   for (const p of allPages) {
-    p.ocr_text = (p.ocr?.data || '').substring(0, 800);
-    delete p.ocr;
     if (!pagesByBook.has(p.book_id)) pagesByBook.set(p.book_id, []);
     pagesByBook.get(p.book_id).push(p);
   }
@@ -265,10 +118,10 @@ for (let i = 0; i < allBooks.length; i += BATCH_SIZE) {
     if (!pages || pages.length === 0) { noData++; continue; }
     checked++;
 
-    // Score all pages
+    // Score all pages using shared scoring module
     const scored = pages
       .map(p => {
-        const { score, reason } = scorePage(p);
+        const { score, reason } = scorePageForCover(p, { bookTitle: book.title });
         return { page: p, score, reason };
       })
       .sort((a, b) => b.score - a.score);
