+++ scripts/workers/pipeline-orchestrator.mjs
@@ -342,128 +342,8 @@ function isDigitizerOcr(ocrData) {
     /ex[\s\-.]?libris|bookplate|library\s+stamp/i.test(start);
 }
 
-/**
- * Score a page as a potential book cover using OCR content analysis.
- * Detects binding photos, digitizer inserts, bookplates (incl. BPH pelican),
- * and prefers decorated title pages with woodcuts/borders/printer's marks.
- *
- * @param {Object} page - Page document with page_number, page_type, ocr.data
- * @returns {{ score: number, reason: string }}
- */
-function scorePageForCover(page) {
-  const ocr = (page.ocr?.data || '').substring(0, 800).toLowerCase();
-  const pageType = page.page_type;
-  const pageNum = page.page_number;
-  let score = 0;
-  let reason = pageType || 'unknown';
-
-  // Skip hidden pages
-  if (page.hidden) return { score: -200, reason: 'hidden' };
-
-  // --- NEGATIVE signals ---
-
-  if (pageType === 'blank') return { score: -100, reason: 'blank' };
-  if (pageType === 'digitizer-notice') return { score: -90, reason: 'digitizer-notice' };
-
-  // Book binding / cover photos
-  if (ocr.includes('front cover') || (ocr.includes('external') && ocr.includes('cover')) ||
-      (ocr.includes('binding') && ocr.includes('spine')) || ocr.includes('fore-edge') ||
-      ocr.includes('closed book') || ocr.includes('book block') ||
-      ocr.includes('leather-bound') || ocr.includes('leather bound') ||
-      ((ocr.includes('blind-stamp') || ocr.includes('blind stamp')) && ocr.includes('cover'))) {
-    return { score: -80, reason: 'physical book photo' };
-  }
-
-  // Digitizer inserts and portal pages
-  if (ocr.includes('digitized by') || (ocr.includes('google') && ocr.includes('logo')) ||
-      (ocr.includes('internet archive') && ocr.includes('logo')) ||
-      ocr.includes('proquest') || ocr.includes('early european books') ||
-      ocr.includes('hathitrust') || ocr.includes('scan sheet') ||
-      ocr.includes('e-rara.ch') || ocr.includes('www.e-rara') ||
-      ocr.includes('gallica.bnf') || ocr.includes('mdz-nbn') ||
-      ocr.includes('digital.staatsbibliothek') || ocr.includes('daten.digitale-sammlungen')) {
-    return { score: -70, reason: 'digitizer insert' };
-  }
-
-  // BPH pelican bookplate
-  const isBphBookplate = (ocr.includes('pelican') && (ocr.includes('piety') || ocr.includes('nest') || ocr.includes('young') || ocr.includes('hermetica'))) ||
-      ocr.includes('philosophia hermetica') || ocr.includes('philosophica hermetica');
-  if (isBphBookplate && pageType !== 'title-page') {
-    return { score: -65, reason: 'BPH pelican bookplate' };
-  }
-
-  // Ex-libris / bookplates (but allow on title pages with small inscriptions)
-  const hasExLibris = ocr.includes('ex-libris') || ocr.includes('exlibris') || ocr.includes('ex libris') ||
-      ocr.includes('bookplate') || (ocr.includes('ownership') && ocr.includes('stamp')) ||
-      ocr.includes('library stamp') || ocr.includes('library sticker');
-  if (hasExLibris && pageType !== 'title-page') {
-    return { score: -60, reason: 'ex-libris/bookplate' };
-  }
-
-  // Bleed-through / ghosting
-  if (ocr.includes('bleed-through') || ocr.includes('ghosting') || ocr.includes('mirrored impression')) {
-    return { score: -50, reason: 'bleed-through' };
-  }
-
-  if (pageType === 'toc' || pageType === 'index') return { score: -20, reason: pageType };
-  if (pageType === 'colophon') return { score: -10, reason: 'colophon' };
-
-  // --- POSITIVE signals ---
-
-  const hasHeadings = (ocr.match(/^#+ .+/gm) || []).length;
-  const isTitlePage = pageType === 'title-page';
-  const looksLikeTitlePage = !pageType && hasHeadings >= 3;
-
-  if (isTitlePage) { score += 80; reason = 'title-page'; }
-  else if (looksLikeTitlePage) { score += 70; reason = 'likely title-page'; }
-
-  if (isTitlePage || looksLikeTitlePage) {
-    // Publisher imprint bonus
-    if (ocr.includes('excudebat') || ocr.includes('typis') || ocr.includes('apud') ||
-        ocr.includes('impensis') || ocr.includes('sumptibus') || ocr.includes('officina') ||
-        ocr.includes('printed by') || ocr.includes('published by') || ocr.includes('printed for')) {
-      score += 15;
-      reason = (isTitlePage ? 'title-page' : 'likely title-page') + ' with imprint';
-    }
-    // Decorative elements bonus
-    if (ocr.includes('decorative') || ocr.includes('ornamental') || ocr.includes('border') ||
-        ocr.includes('frame') || ocr.includes('vignette') || ocr.includes('woodcut') ||
-        ocr.includes('architectural') || ocr.includes("printer's mark") ||
-        ocr.includes("printer's device")) {
-      score += 20;
-      reason = 'decorated title-page';
-    }
-  }
-
-  // Frontispieces — but not binding photos mislabeled as frontispiece
-  if (pageType === 'frontispiece') {
-    if (ocr.includes('cover') && (ocr.includes('leather') || ocr.includes('binding') || ocr.includes('marbled'))) {
-      return { score: -80, reason: 'binding photo (mislabeled frontispiece)' };
-    }
-    score += 90; reason = 'frontispiece';
-    if (ocr.includes('engrav') || ocr.includes('allegor') || ocr.includes('portrait') ||
-        ocr.includes('emblem') || ocr.includes('woodcut')) {
-      score += 15; reason = 'frontispiece with engraving';
-    }
-  }
-
-  if (pageType === 'illustration') {
-    if (ocr.includes('photograph') && (ocr.includes('fore-edge') || (ocr.includes('book') && ocr.includes('closed')))) {
-      return { score: -80, reason: 'physical book photo' };
-    }
-    score += 60; reason = 'illustration';
-  }
-
-  if (pageType === 'dedication') { score += 15; reason = 'dedication'; }
-  if (pageType === 'text' && score === 0) { score += 5; reason = 'text'; }
-
-  // Position bonus
-  if (pageNum <= 5) score += 5;
-  else if (pageNum <= 10) score += 3;
-  else if (pageNum > 15) score -= 3;
-
-  return { score, reason };
-}
+// Cover scoring — imported from shared module
+import { scorePageForCover } from '../lib/cover-scoring.mjs';
 
 /**
  * Select the best cover page for a book from its first N pages.
@@ -474,7 +354,7 @@ function scorePageForCover(page) {
  * @param {number} maxPages - Max pages to consider (default 20)
  * @returns {Promise<{page: Object, score: number, reason: string} | null>}
  */
-async function selectBestCoverPage(db, bookId, maxPages = 20) {
+async function selectBestCoverPage(db, bookId, maxPages = 20, bookTitle = null) {
   const pages = await db.collection('pages').find(
     { book_id: bookId, page_number: { $lte: maxPages } },
     { projection: {
@@ -487,7 +367,7 @@ async function selectBestCoverPage(db, bookId, maxPages = 20) {
   if (!pages.length) return null;
 
   const scored = pages
-    .map(p => ({ page: p, ...scorePageForCover(p) }))
+    .map(p => ({ page: p, ...scorePageForCover(p, { bookTitle }) }))
     .sort((a, b) => b.score - a.score);
 
   const best = scored[0];
@@ -3207,7 +3087,7 @@ Rules:
               // cover immediately instead of waiting for the full pipeline to complete.
               if (book.thumbnail_source !== 'manual') {
                 try {
-                  const coverResult = await selectBestCoverPage(db, book.id);
+                  const coverResult = await selectBestCoverPage(db, book.id, 20, book.title);
                   if (coverResult) {
                     await db.collection('books').updateOne(
                       { id: book.id },
@@ -4271,7 +4151,7 @@ Rules:
           }
 
           // Smart OCR-based cover selection
-          const coverResult = await selectBestCoverPage(db, book.id);
+          const coverResult = await selectBestCoverPage(db, book.id, 20, book.title);
           if (coverResult && coverResult.url !== book.thumbnail) {
             await db.collection('books').updateOne(
               { id: book.id },
